final_changes

+++ app/controllers/todos_controller.rb
@@ -6,6 +6,8 @@ class TodosController < ApplicationController
 
     # GET /todo_lists/:todos_list_id/todos
   def index
+    authorize @todo_list, :show?
+
     todos = @todo_list.todos
     render json: todos, status: :ok
   end
@@ -13,6 +15,9 @@ class TodosController < ApplicationController
     # POST /todo_list/:todo_list_id/todos
   def create
     todo = @todo_list.todos.build(todo_params)
+
+    authorize todo
+
     if todo.save
       render json: todo, status: :created
     else
@@ -20,25 +25,29 @@ class TodosController < ApplicationController
     end
   end
 
+
   #Get /todo_lists/:todo_list_id/todos/:id
   def show
+    authorize @todo
     render json: @todo, status: :ok
   end
 
   # PUT /todo_list/:todo_list_id/todos/:id
   def update
-  if @todo.update(todo_params)
-    render json: @todo, status: :ok
-  else
-    render json: { errors: @todo.errors.full_messages }, status: :unprocessable_entity
-   end
+    authorize @todo
+
+    if @todo.update(todo_params)
+      render json: @todo, status: :ok
+    else
+      render json: { errors: @todo.errors.full_messages }, status: :unprocessable_entity
+    end
   end
 
   # DELETE  by /todo_list/:todo_list_id/todos/:id
   def destroy
-    
+    authorize @todo
+
     if @todo.destroy
-       puts "Updated Todo: #{@todo.inspect}"
       render json: { message: 'Todo deleted successfully' }, status: :ok
     else
       render json: { errors: @todo.errors.full_messages }, status: :unprocessable_entity
@@ -47,11 +56,24 @@ class TodosController < ApplicationController
 
   private
   
-  def set_todo_list
-    @todo_list = current_user.todo_lists.find_by(id: params[:todo_list_id])
-    unless @todo_list
-      render json: { error: "Todo list not found" }, status: :not_found 
-    end
+ def set_todo_list
+  # Fetch the todo_list either for the current user or if they are a collaborator
+  @todo_list = TodoList.find_by(id: params[:todo_list_id])
+  
+  # Check if the current user is either the owner or a collaborator of the todo_list
+  unless @todo_list && (user_is_owner? || collaborator?)
+    render json: { error: "Todo list not found or you don't have permission" }, status: :not_found
+  end
+ end
+
+  # Check if the current user is the owner of the todo list
+  def user_is_owner?
+    @todo_list.user == current_user
+  end
+
+  # Check if the current user is a collaborator on the todo list
+  def collaborator?
+    @todo_list.collaborators.exists?(user_id: current_user.id)
   end
 
   def set_todo
